pyproject.toml
[project]
name = "somnus"
version = "2.0.0"
description = "Discord bot to remotely control and manage multiple Minecraft servers"
authors = [
    { name = "2mal3", email = "56305732+2mal3@users.noreply.github.com" },
    { name = "programmer-44", email = "github@keinbock.mozmail.com" },
]
dependencies = [
    "discord-py==2.*",
    "python-dotenv==1.*",
    "wakeonlan==3.*",
    "ping3==4.*",
    "pexpect==4.*",
    "mcstatus==11.*",
    "pydantic==2.*",
    "aiofiles==24.*",
    "yet-another-i18n==0.2.*",
    "toml==0.10.*",
    "asyncer==0.0.17",
]
readme = "README.md"
requires-python = ">= 3.11"

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
# Rules are explained here: https://docs.astral.sh/ruff/rules
select = ["ALL"]
ignore = [
    "FBT",
    "D",    # we are not a library
    "PT",
    "B",
    "EM",
    "TRY",
    "PGH003",
    "COM812",   # unnecessary complex
    "TCH",
    "I",    # way to complex to automate this
    "PTH",
    "RET",
    "BLE",
    "ARG",  # discord.py needs some arguments to be present even when we don't use them
    "RUF005",   # Bad readability / unnecessary complex
    "DTZ005",
    "UP",
    "PD",   # we are not data scientists
    "S101", # we need assert for tests
    "S110", # empty Except
    "SIM105", # empty Except without else
]

[tool.rye.scripts]
dev = "python3 -m somnus.discord_provider.main"
prod = "python3 -m somnus.__main__"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
virtual = true
dev-dependencies = [
    "podman==5.*",
    "pytest==8.*",
    "ruff==0.15.11",
    "ty==0.0.31",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["somnus"]
